requirements.txt
discord.py
aiohttp
sqlite-utils
python-dotenv
black              
isort              
pytest      
ollama      
